first commit
game could do with refactoring and things being put into classes but the game works. Game desc. at the time of this commit: can shoot 2 bullets at any time but then have to wait for bullets to hit the enemy or to escape display before being able to fire again. can move left and right. cant go outside of the display. There is a score and various game elements change dependent on the score. Main menu has been implemented and gets resorted to at the end of the game with the score achieved